requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-environ==0.11.2
djangorestframework==3.14.0
idna==3.4
psycopg2==2.9.7
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
whitenoise==6.5.0
